More bug fixes

+++ bot.py
@@ -24,23 +24,27 @@ async def on_raw_reaction_add(payload):
 
 @bot.event
 async def on_command_error(ctx, error):
-    await ctx.message.delete()
-    if(isinstance(error, discord.ext.commands.errors.MissingRequiredArgument)):
-        await ctx.message.author.send("Enter a user!")
-    elif(isinstance(error, discord.errors.Forbidden)):
-        await ctx.message.author.send("Type your command into a valid text channel!")
+    if(ctx.guild):
+        await ctx.message.delete()
+        if(isinstance(error, discord.ext.commands.errors.MissingRequiredArgument)):
+            await ctx.message.author.send("Enter a user!")
+    else:
+        await ctx.send("Type your command into a valid text channel!")
 
 
-@bot.command(name="dm", brief="Anonymously dms user")
+@bot.command(name="dm", brief="Anonymously dms user (anon-dm @recipient your sentence)")
 async def dm(ctx, user: discord.User, *, message=None):
-    if(message is None):
-        await ctx.message.author.send("Enter a message to send to " + user.display_name)
-        return
-    successMsg = await ctx.message.author.send("Successfully sent **" + user.display_name + ":** \"" + message + "\"")
-    await successMsg.add_reaction("\U0001F5D1")
-    msg = await user.send("**An anonymous user sent you this:** "+message)
-    await msg.add_reaction("\U0001F5D1")
-    await ctx.message.delete()
+    if(ctx.guild):
+        await ctx.message.delete()
+        if(message is None):
+            await ctx.message.author.send("Enter a message to send to " + user.display_name)
+            return
+        successMsg = await ctx.message.author.send("Successfully sent **" + user.display_name + ":** \"" + message + "\"")
+        await successMsg.add_reaction("\U0001F5D1")
+        msg = await user.send("**An anonymous user sent you this:** "+message)
+        await msg.add_reaction("\U0001F5D1")
+    else:
+        await ctx.send("Type your command into a valid text channel!")
 
 load_dotenv()
 bot.run(getenv('TOKEN'))
